Hide incompatible functions behind build tags
