+++ futures.go
@@ -1,3 +1,5 @@
+//go:build cryptsetup2.4
+
 package cryptsetup
 
 // #cgo pkg-config: libcryptsetup
@@ -8,6 +10,19 @@ import "C"
 import (
 	"unsafe"
 )
+
+// DumpJSON returns JSON-formatted information about a LUKS2 device.
+// C equivalent: crypt_dump_json
+func (device *Device) DumpJSON() (string, error) {
+	cStr := C.CString("")
+	defer C.free(unsafe.Pointer(cStr))
+
+	// crypt_dump_json does not support flags currently, but they are reserved for future use.
+	if res := C.crypt_dump_json(device.cryptDevice, &cStr, C.uint32_t(0)); res != 0 {
+		return "", &Error{functionName: "crypt_dump_json", code: int(res)}
+	}
+	return C.GoString(cStr), nil
+}
 
 // TokenExternalDisable disables external token handlers (plugins) support.
 // If disabled, it cannot be enabled again.
